[modify] gws discussion: permit_comment, permanently options

+++ app/controllers/gws/discussion/comments_controller.rb
@@ -53,14 +53,8 @@ class Gws::Discussion::CommentsController < ApplicationController
     redirect_to action: :index
   end
 
-  def update
-    @item.attributes = get_params
-    @item.in_updated = params[:_updated] if @item.respond_to?(:in_updated)
-    raise "403" unless @item.allowed?(:edit, @cur_user, site: @cur_site)
-    render_update @item.update, location: { action: :index }
-  end
-
   def edit
+    raise "403" if @topic.permanently?
     raise "403" unless @item.allowed?(:edit, @cur_user, site: @cur_site, grants_none_to_owner: true)
     if @item.is_a?(Cms::Addon::EditLock)
       unless @item.acquire_lock
@@ -74,36 +68,25 @@ class Gws::Discussion::CommentsController < ApplicationController
   def update
     @item.attributes = get_params
     @item.in_updated = params[:_updated] if @item.respond_to?(:in_updated)
+    raise "403" if @topic.permanently?
     raise "403" unless @item.allowed?(:edit, @cur_user, site: @cur_site, grants_none_to_owner: true)
     render_update @item.update
   end
 
   def delete
+    raise "403" if @topic.permanently?
     raise "403" unless @item.allowed?(:delete, @cur_user, site: @cur_site, grants_none_to_owner: true)
     render
   end
 
   def destroy
+    raise "403" if @topic.permanently?
     raise "403" unless @item.allowed?(:delete, @cur_user, site: @cur_site, grants_none_to_owner: true)
     render_destroy @item.destroy
   end
 
-  def destroy_all
-    entries = @items.entries
-    @items = []
-
-    entries.each do |item|
-      if item.allowed?(:delete, @cur_user, site: @cur_site, grants_none_to_owner: true)
-        next if item.destroy
-      else
-        item.errors.add :base, :auth_error
-      end
-      @items << item
-    end
-    render_destroy_all(entries.size != @items.size)
-  end
-
   def reply
+    raise "403" unless @topic.permit_comment?
     raise "403" unless @model.allowed?(:edit, @cur_user, site: @cur_site)
     @comment = @model.new get_params
